requirements.txt
filterpy==1.4.5
numpy==2.2.0
opencv_contrib_python==4.10.0.84
PyYAML==6.0.2
